+++ testing/unit_tests_to_autoload/testCases_constraints_str.py
@@ -8,7 +8,7 @@ import jk_jsonmodel
 
 
 
-DATA = jk_jsonmodel.loadModel("""{
+DATA = jk_jsonmodel.loadModelFromStr("""{
 	"aStr": "abc",
 }""")